+++ vkconfig_core/test/test_environment.cpp
@@ -33,9 +33,9 @@ TEST(test_environment, custom_path_no_duplicate) {
     environment.Reset(Environment::DEFAULT);
 
     EXPECT_EQ(true, environment.AppendCustomLayerPath("./path"));
-    EXPECT_EQ(1, environment.GetCustomLayerPaths().size());
+    EXPECT_EQ(1, environment.GetUserDefinedLayersPaths(USER_DEFINED_LAYERS_PATHS_GUI).size());
     EXPECT_EQ(false, environment.AppendCustomLayerPath("./path"));
-    EXPECT_EQ(1, environment.GetCustomLayerPaths().size());
+    EXPECT_EQ(1, environment.GetUserDefinedLayersPaths(USER_DEFINED_LAYERS_PATHS_GUI).size());
 
     environment.Reset(Environment::SYSTEM);  // Don't change the system settings
 }
@@ -47,11 +47,11 @@ TEST(test_environment, custom_path_not_found) {
     environment.Reset(Environment::DEFAULT);
 
     EXPECT_EQ(false, environment.RemoveCustomLayerPath("./path"));
-    EXPECT_EQ(0, environment.GetCustomLayerPaths().size());
+    EXPECT_EQ(0, environment.GetUserDefinedLayersPaths(USER_DEFINED_LAYERS_PATHS_GUI).size());
     EXPECT_EQ(true, environment.AppendCustomLayerPath("./path"));
-    EXPECT_EQ(1, environment.GetCustomLayerPaths().size());
+    EXPECT_EQ(1, environment.GetUserDefinedLayersPaths(USER_DEFINED_LAYERS_PATHS_GUI).size());
     EXPECT_EQ(true, environment.RemoveCustomLayerPath("./path"));
-    EXPECT_EQ(0, environment.GetCustomLayerPaths().size());
+    EXPECT_EQ(0, environment.GetUserDefinedLayersPaths(USER_DEFINED_LAYERS_PATHS_GUI).size());
 
     environment.Reset(Environment::SYSTEM);  // Don't change the system settings
 }