+++ vkconfig/CHANGELOG.md
@@ -25,6 +25,10 @@
 ### Features:
 - Add *Synchronization* default configuration #1304
 
+### Improvements:
+- Add display of VK_LAYER_PATH content in Vulkan Development Status
+- Renamed "Custom Path" by "User-Defined Path"
+
 ## [Vulkan Configurator 2.1.0 for Vulkan SDK 1.2.162.1](https://github.com/LunarG/VulkanTools/tree/sdk-1.2.162.1) - January 2021
 
 ### Features: